# Speed up dataset opening by caching many HDF5 variables