We are trying to use h5netcdf in our pipeline since it offers alot of flexibility in how HDF5 is accessed and used.

However, we noticed that for some of our datasets, they open with 
* xarray + netcdf4-python in about 80 ms
* xarray + h5netcdf: 600 ms.

I've been working on speeding things up, much of it involces caching results from accesses to the HDF5 file.

There are a few questions I had, so I'm opening up this merge request to give me a place to ask them.

I've narrowed the gap to 240 ms vs 80 ms.  I'll also try to "donate" a dataset. It mostly has 130 data variables, that share many coordinates. Our new datasets really cut down on the coordinates (which I think helps) but given the proliferation of these datasets, even internally, I'm interested in general speedups where I can get them.

As promised, the dataset one can use to benchmark:

The dataset contains `images` (I made them monochrome for easy viewing) which can be viewed.

This is a screenshot from our viewer, but you can likely open them in matplotlib too.
```python
import xarray as xr
import h5netcdf  # pay the cost of lazy loading here so it doesn't appear in the benchmarks
filename = 'sample_file_zebrafish_0.18.55_4uCams_small.nc'
dataset = xr.open_dataset(filename, engine='h5netcdf')
```
```python
# Visualize
from matplotlib import pyplot as plt
# There are 4 images (0, 0), (0, 1), (1, 0), and (1, 1)
plt.imshow(dataset['images'][0, 0], cmap='gray')
```

![image](https://user-images.githubusercontent.com/90008/202945018-5ffa9290-3dde-4544-8e0d-5ecce4c5fe66.png)


[sample_file_zebrafish_0.18.55_4uCams_small.zip](https://github.com/h5netcdf/h5netcdf/files/10051903/sample_file_zebrafish_0.18.55_4uCams_small.zip)

Initial timing on my machine shows that this takes about 
* 250 ms to load with the main branch.
* With this branch, it cuts it down to 125 ms (but at the time of adding the dataset, it has a segfault)
* xarray + netcdf take about 31 ms to load the file.

The benchmark is the following:
```python
import xarray as xr
import h5netcdf  # eagerly import for benchmark
import netCDF4  # eagerly import for benchmark
filename = 'sample_file_zebrafish_0.18.55_4uCams_small.nc'

%%timeit
xr.open_dataset(filename, engine="netcdf4").close()
# %%
%%timeit
xr.open_dataset(filename, engine="h5netcdf").close()
```


- [ ] Closes Issue #xxxx
- [ ] Tests added
- [ ] Changes are documented in `CHANGELOG.rst`