@kmuehlbauer 2022-11-21
@hmaarrfk Thanks for raising this and also for providing use-case data and MCVE. Performance improvements are most wanted. 

I'll need to dive into this, so please bear with me if this take some time.

My initial thoughts:

- dataset/name caching can be done when opened for read-only in any case, for write/append we would need to think how to apply caching/refresh cache on write
- any improvements by using h5py-machinery (eg. our dereferencing example) should be implemented

One question, though: Where does your example segfault? I could not get any segfault.