@hmaarrfk 2022-11-21
> One question, though: Where does your example segfault? I could not get any segfault.

Running `%%timeit`. I mean, it stopped happenning after a while too, so 

> dataset/name caching can be done when opened for read-only in any case, for write/append we would need to think how to apply caching/refresh cache on write

This was my fear.

I think there are some "clear improvements" here that I can break out of this mega PR that don't change anything outside of a single function.