@kmuehlbauer 2022-11-21
> I think there are some "clear improvements" here that I can break out of this mega PR that don't change anything outside of a single function.

Yes, please!