@hmaarrfk 2022-11-29
Sigh, there are some marginal gains to be made, but ultimately, alot of stuff relies on structures like:

```
self._h5ds
```

these seem to be "expensive" to create. The code here really relies on just creating the `self._h5ds` structure once. 

That said, I think I got mixed results in my benchmarks. so I'm not super excited, but yeah, I haven't given up.